fix warn for clang++

+++ db/compaction/compaction_job.cc
@@ -828,7 +828,7 @@ try {
   rpc_params.db_session_id = this->db_session_id_;
   rpc_params.full_history_ts_low = this->full_history_ts_low_;
 //rpc_params.compaction_job_stats = this->compaction_job_stats_;
-  rpc_params.max_subcompactions = num_threads;
+  rpc_params.max_subcompactions = uint32_t(num_threads);
 
   const uint64_t start_micros = env_->NowMicros();
   auto exec_factory = imm_cfo->compaction_executor_factory.get();
